// frontend/src/context/types/notification.ts
export enum GetNotificationsType {
	NOTIFICATIONS_LIST_REQUEST = 'NOTIFICATIONS_LIST_REQUEST',
	NOTIFICATIONS_LIST_SUCCESS = 'NOTIFICATIONS_LIST_SUCCESS',
	NOTIFICATIONS_LIST_FAIL = 'NOTIFICATIONS_LIST_FAIL',
}
export enum MarkAsReadType {
	MARK_AS_READ_REQUEST = 'MARK_AS_READ_REQUEST',
	MARK_AS_READ_SUCCESS = 'MARK_AS_READ_SUCCESS',
	MARK_AS_READ_FAIL = 'MARK_AS_READ_FAIL',
}
export enum MarkAllAsReadType {
	MARK_ALL_AS_READ_REQUEST = 'MARK_ALL_AS_READ_REQUEST',
	MARK_ALL_AS_READ_SUCCESS = 'MARK_ALL_AS_READ_SUCCESS',
	MARK_ALL_AS_READ_FAIL = 'MARK_ALL_AS_READ_FAIL',
}